Factored out the BifrostReactWidget into OnboardingWidget and VisualizationScreen
