+++ src/components/BifrostReactWidget.tsx
@@ -1,40 +1,16 @@
 /** @jsx jsx */
 import { jsx, css, ThemeProvider, Global } from '@emotion/react';
 
-// import Graph from './Graph';
-import Sidebar from './Sidebar/Sidebar';
+import { useState } from 'react';
 import { WidgetModel } from '@jupyter-widgets/base';
 import {
-  BifrostModelContext,
   useModelState,
   Flags,
-  SuggestedGraphs,
-  GraphData,
-  QuerySpec,
+  BifrostModelContext,
 } from '../hooks/bifrost-model';
-import React, { useState } from 'react';
-import ChartChooser from './Onboarding/ChartChooser';
-import NavBar from './NavBar';
-import ColumnScreen from './Onboarding/ColumnScreen';
-import { VisualizationSpec } from 'react-vega';
-import Graph from './Graph';
 import theme from '../theme';
-
-import { build } from 'compassql/build/src/schema';
-import { recommend } from 'compassql/build/src/recommend';
-import { mapLeaves } from 'compassql/build/src/result';
-import { SpecQueryModel } from 'compassql/build/src/model';
-import { FieldQuery } from 'compassql/build/src/query/encoding';
-
-const bifrostWidgetCss = css`
-  // Element-based styles
-  //===========================================================
-  display: grid;
-  grid-template-columns: 2fr 1fr;
-  grid-template-rows: auto 1fr;
-  grid-template-areas: 'nav sidebar' 'graph sidebar';
-  max-width: calc(100% - 64px);
-`;
+import OnboardingWidget from './Onboarding/OnboardingWidget';
+import VisualizationScreen from './VisualizationScreen';
 
 const globalStyles = (theme: any) => css`
   // Global styles for the widget
@@ -63,6 +39,7 @@ const globalStyles = (theme: any) => css`
       background: transparent;
       margin: 0;
       padding: 0;
+      margin-right: 15px;
       color: initial;
     }
   }
@@ -88,123 +65,33 @@ export default function BifrostReactWidget(props: BifrostReactWidgetProps) {
   return (
     <ThemeProvider theme={theme}>
       <BifrostModelContext.Provider value={props.model}>
-        <OnBoardingWidget />
         <Global styles={globalStyles} />
+        <BifrostReactWidgetDisplay />
       </BifrostModelContext.Provider>
     </ThemeProvider>
   );
 }
 
-export function OnBoardingWidget() {
+function BifrostReactWidgetDisplay() {
   const flags = useModelState<Flags>('flags')[0];
-  const setSelectedSpec = useState<VisualizationSpec>({})[1];
-  const [suggestedGraphs, setSuggestedGraphs] =
-    useModelState<SuggestedGraphs>('suggested_graphs');
-  const querySpec = useModelState<QuerySpec>('query_spec')[0];
-  const data = useModelState<GraphData>('graph_data')[0];
-  const columnChoices = useModelState<string[]>('df_columns')[0];
 
-  const [screenName, setScreenName] = flags['columns_provided']
-    ? flags['kind_provided']
-      ? useState('straight_visualize')
-      : useState('chartChooser')
-    : useState('columnChooser');
-
-  let Screen: JSX.Element;
-  const preSelectedColumns = new Set<string>();
-
-  switch (screenName) {
-    case 'columnChooser':
-      querySpec.spec.encodings.forEach((encoding: FieldQuery) => {
-        if (
-          encoding.channel !== '?' &&
-          columnChoices.includes(encoding.field as string)
-        ) {
-          preSelectedColumns.add(encoding.field as string);
-        }
-      });
-
-      Screen = (
-        <ColumnScreen
-          onNext={() => setScreenName('chartChooser')}
-          preSelectedColumns={preSelectedColumns}
-        />
-      );
-      break;
-    case 'chartChooser':
-      Screen = (
-        <ChartChooser
-          onChartSelected={(spec) => {
-            setSelectedSpec(spec);
-            setScreenName('visualize');
-          }}
-          onBack={
-            suggestedGraphs.length === 0
-              ? undefined
-              : () => setScreenName('columnChooser')
-          }
-        />
-      );
-
-      if (suggestedGraphs.length === 0) {
-        const opt = {};
-
-        const schema = build(data, opt);
-
-        const result = recommend(querySpec, schema, opt).result;
-
-        const vlTree = mapLeaves(result, (item: SpecQueryModel) => {
-          const newSpec: Record<string, any> = item.toSpec();
-          newSpec['params'] = (querySpec.spec as Record<string, any>)['params'];
-          return newSpec;
-        });
-
-        const items = vlTree.items;
-
-        if (items.length !== 0) {
-          setSuggestedGraphs(items as SuggestedGraphs);
-        }
-      }
-      break;
-    case 'visualize':
-      Screen = (
-        <VisualizationScreen onPrevious={() => setScreenName('chartChooser')} />
-      );
-      break;
-    default:
-      Screen = <VisualizationScreen />;
-      break;
-  }
-  return Screen;
-}
-
-function VisualizationScreen({ onPrevious }: { onPrevious?: () => void }) {
+  const [screenName, setScreenName] = useState<string>(
+    !flags['columns_provided']
+      ? 'columnChooser'
+      : !flags['kind_provided']
+      ? 'chartChooser'
+      : 'straight_visualize'
+  );
   return (
-    <article className="BifrostWidget" css={bifrostWidgetCss}>
-      {onPrevious ? (
-        <GridArea area="nav">
-          <NavBar onBack={onPrevious} />
-        </GridArea>
-      ) : null}
-      <GridArea area="graph">
-        {onPrevious ? <Graph onBack={onPrevious} /> : <Graph />}
-      </GridArea>
-      <GridArea area="sidebar">
-        <Sidebar />
-      </GridArea>
-    </article>
+    <div className="bifrost-widget-display">
+      {screenName === 'straight_visualize' ? (
+        <VisualizationScreen />
+      ) : (
+        <OnboardingWidget
+          screenName={screenName}
+          setScreenName={setScreenName}
+        />
+      )}
+    </div>
   );
 }
-
-interface GridAreaProps {
-  area: string;
-  children?: any;
-}
-
-const GridArea = React.forwardRef<HTMLDivElement, GridAreaProps>(
-  (props, ref) => (
-    <div style={{ gridArea: props.area }} ref={ref}>
-      {props.children}
-    </div>
-  )
-);
